kickstart-metadata: initial entry
Example entry for the 'Kickstart Metadata' field in the beaker
provisioning interface.

This is still not complete, because it does not create separate
partitions for /home and /work, but at least the disk is partitioned
in a way that allows for /work to be created and for it to share all
the space with /home under the /.

So, we only get the / partition with this, and /home and /work, though
/home is created and shares the space with / and anything else created
uner /.

Also, /boot is still coming up as xfs, when I would like that also to
be ext4, but good enough for now.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>